Fixed crashing bug when parsing invalid messages
Fixes a crash that was observed when starting to parse a TLS encrypted stream while assuming to parse plain MQTT data.

So what happened? The TLS connect message that arrived at looked like this:

```
<Buffer 16 03 01 01 01 01 00 00 fd 03 03 00 09 9a 85 fc 19 19 fa 06 a4
8c a5 53 5d 6e c1 01 23 be 68 00 85 b9 0c 29 8b 24 87 30 09 e6 e4 00 00
80 c0 2f c0 2b ... >
```

Note the `16` at the start, this suggests that we have a regular CONNECT packet. The next byte 3 suggests a remaining length of the packet of 3 bytes which is obviously not true, but the implementation of the parser just checks if at least 3 bytes remain in the buffer before it starts processing the payload.

The parseConnect function of the parser now checks the next two bytes `01 01` which would describe the protocol identifier string length. Usually
this would be 4 or 6 bytes, but here we have 257 bytes. So the next 257 bytes are consumed into a string and our parsing pointer is at buffer position 259. Unfortunately, the buffer only has 260 bytes in total. That is the next parsing steps produce an out of range exception..

We obviously have various problems here:

- the parser does not check for plausibility when a string ought to be longer than the packet
- the parser does not check for a valid protocol identifier string at all
- the parser does not check (at some points) if the buffer has remaining length before parsing the next bytes
- the parser does not catch any (further, unexpected) exceptions in order to safe the upper layers from running into bad behavior